Use a normal handler (wtf was I thinking?)
